Move RE forms to folder
