+++ src/module/item/sheet/base.ts
@@ -95,7 +95,7 @@ export class ItemSheetPF2e<TItem extends ItemPF2e> extends ItemSheet<TItem> {
         this.ruleElementForms = {};
         for (const [idx, rule] of rules.entries()) {
             const FormClass = RULE_ELEMENT_FORMS[String(rule.key)] ?? RuleElementForm;
-            this.ruleElementForms[Number(idx)] = new FormClass(idx, rule);
+            this.ruleElementForms[Number(idx)] = new FormClass(this, idx, rule);
         }
 
         return {
